feat(debug-tab): add Debug main tab with Console/Network sub-tabs
Replace the "Issue" main tab with a "Debug" grouping tab containing
three sub-tabs: Issue (existing IssueTab), Console, and Network.
Extract viewer content from ConsoleLogPreviewDialog and
NetworkLogPreviewDialog into standalone ConsoleLogContent and
NetworkLogContent components for reuse. Add periodic sync polling
(1.5s) for live log updates when Console/Network sub-tabs are active.
