+++ src/sidepanel/tabs/DebugTab.tsx
@@ -1,0 +1,50 @@
+import { useState } from "react";
+import { useT } from "@/i18n";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { IssueTab } from "./IssueTab";
+import { ConsoleSubTab } from "./ConsoleSubTab";
+import { NetworkSubTab } from "./NetworkSubTab";
+
+type DebugSubTab = "issue" | "console" | "network";
+
+export function DebugTab() {
+  const t = useT();
+  const [sub, setSub] = useState<DebugSubTab>("issue");
+
+  return (
+    <Tabs
+      value={sub}
+      onValueChange={(v) => setSub(v as DebugSubTab)}
+      className="flex min-h-0 flex-1 flex-col gap-0"
+    >
+      <div className="shrink-0 border-b border-border px-4 py-4">
+        <TabsList className="grid h-9 w-full grid-cols-3">
+          <TabsTrigger value="issue">{t("debug.tab.issue")}</TabsTrigger>
+          <TabsTrigger value="console">{t("debug.tab.console")}</TabsTrigger>
+          <TabsTrigger value="network">{t("debug.tab.network")}</TabsTrigger>
+        </TabsList>
+      </div>
+
+      <TabsContent
+        value="issue"
+        className="mt-0 flex min-h-0 flex-1 flex-col data-[state=inactive]:hidden"
+      >
+        <IssueTab />
+      </TabsContent>
+
+      <TabsContent
+        value="console"
+        className="mt-0 flex min-h-0 flex-1 flex-col data-[state=inactive]:hidden"
+      >
+        <ConsoleSubTab active={sub === "console"} />
+      </TabsContent>
+
+      <TabsContent
+        value="network"
+        className="mt-0 flex min-h-0 flex-1 flex-col data-[state=inactive]:hidden"
+      >
+        <NetworkSubTab active={sub === "network"} />
+      </TabsContent>
+    </Tabs>
+  );
+}
